requirement.txt
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
aws-psycopg2==1.3.8
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
boto3==1.24.81
botocore==1.27.81
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.8.0
fonttools==4.33.3
gTTS==2.2.4
idna==3.3
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pipenv==2022.10.12
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
psycopg-binary==3.1.2
psycopg2-binary==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyobjc==8.5
pyobjc-core==8.5
pyobjc-framework-Accessibility==8.5
pyobjc-framework-Accounts==8.5
pyobjc-framework-AddressBook==8.5
pyobjc-framework-AdServices==8.5
pyobjc-framework-AdSupport==8.5
pyobjc-framework-AppleScriptKit==8.5
pyobjc-framework-AppleScriptObjC==8.5
pyobjc-framework-ApplicationServices==8.5
pyobjc-framework-AppTrackingTransparency==8.5
pyobjc-framework-AudioVideoBridging==8.5
pyobjc-framework-AuthenticationServices==8.5
pyobjc-framework-AutomaticAssessmentConfiguration==8.5
pyobjc-framework-Automator==8.5
pyobjc-framework-AVFoundation==8.5
pyobjc-framework-AVKit==8.5
pyobjc-framework-BusinessChat==8.5
pyobjc-framework-CalendarStore==8.5
pyobjc-framework-CallKit==8.5
pyobjc-framework-CFNetwork==8.5
pyobjc-framework-ClassKit==8.5
pyobjc-framework-CloudKit==8.5
pyobjc-framework-Cocoa==8.5
pyobjc-framework-Collaboration==8.5
pyobjc-framework-ColorSync==8.5
pyobjc-framework-Contacts==8.5
pyobjc-framework-ContactsUI==8.5
pyobjc-framework-CoreAudio==8.5
pyobjc-framework-CoreAudioKit==8.5
pyobjc-framework-CoreBluetooth==8.5
pyobjc-framework-CoreData==8.5
pyobjc-framework-CoreHaptics==8.5
pyobjc-framework-CoreLocation==8.5
pyobjc-framework-CoreMedia==8.5
pyobjc-framework-CoreMediaIO==8.5
pyobjc-framework-CoreMIDI==8.5
pyobjc-framework-CoreML==8.5
pyobjc-framework-CoreMotion==8.5
pyobjc-framework-CoreServices==8.5
pyobjc-framework-CoreSpotlight==8.5
pyobjc-framework-CoreText==8.5
pyobjc-framework-CoreWLAN==8.5
pyobjc-framework-CryptoTokenKit==8.5
pyobjc-framework-DataDetection==8.5
pyobjc-framework-DeviceCheck==8.5
pyobjc-framework-DictionaryServices==8.5
pyobjc-framework-DiscRecording==8.5
pyobjc-framework-DiscRecordingUI==8.5
pyobjc-framework-DiskArbitration==8.5
pyobjc-framework-DVDPlayback==8.5
pyobjc-framework-EventKit==8.5
pyobjc-framework-ExceptionHandling==8.5
pyobjc-framework-ExecutionPolicy==8.5
pyobjc-framework-ExternalAccessory==8.5
pyobjc-framework-FileProvider==8.5
pyobjc-framework-FileProviderUI==8.5
pyobjc-framework-FinderSync==8.5
pyobjc-framework-FSEvents==8.5
pyobjc-framework-GameCenter==8.5
pyobjc-framework-GameController==8.5
pyobjc-framework-GameKit==8.5
pyobjc-framework-GameplayKit==8.5
pyobjc-framework-ImageCaptureCore==8.5
pyobjc-framework-IMServicePlugIn==8.5
pyobjc-framework-InputMethodKit==8.5
pyobjc-framework-InstallerPlugins==8.5
pyobjc-framework-InstantMessage==8.5
pyobjc-framework-Intents==8.5
pyobjc-framework-IntentsUI==8.5
pyobjc-framework-IOSurface==8.5
pyobjc-framework-iTunesLibrary==8.5
pyobjc-framework-KernelManagement==8.5
pyobjc-framework-LatentSemanticMapping==8.5
pyobjc-framework-LaunchServices==8.5
pyobjc-framework-libdispatch==8.5
pyobjc-framework-LinkPresentation==8.5
pyobjc-framework-LocalAuthentication==8.5
pyobjc-framework-LocalAuthenticationEmbeddedUI==8.5
pyobjc-framework-MailKit==8.5
pyobjc-framework-MapKit==8.5
pyobjc-framework-MediaAccessibility==8.5
pyobjc-framework-MediaLibrary==8.5
pyobjc-framework-MediaPlayer==8.5
pyobjc-framework-MediaToolbox==8.5
pyobjc-framework-Metal==8.5
pyobjc-framework-MetalKit==8.5
pyobjc-framework-MetalPerformanceShaders==8.5
pyobjc-framework-MetalPerformanceShadersGraph==8.5
pyobjc-framework-MetricKit==8.5
pyobjc-framework-MLCompute==8.5
pyobjc-framework-ModelIO==8.5
pyobjc-framework-MultipeerConnectivity==8.5
pyobjc-framework-NaturalLanguage==8.5
pyobjc-framework-NetFS==8.5
pyobjc-framework-Network==8.5
pyobjc-framework-NetworkExtension==8.5
pyobjc-framework-NotificationCenter==8.5
pyobjc-framework-OpenDirectory==8.5
pyobjc-framework-OSAKit==8.5
pyobjc-framework-OSLog==8.5
pyobjc-framework-PassKit==8.5
pyobjc-framework-PencilKit==8.5
pyobjc-framework-Photos==8.5
pyobjc-framework-PhotosUI==8.5
pyobjc-framework-PreferencePanes==8.5
pyobjc-framework-PushKit==8.5
pyobjc-framework-Quartz==8.5
pyobjc-framework-QuickLookThumbnailing==8.5
pyobjc-framework-ReplayKit==8.5
pyobjc-framework-SafariServices==8.5
pyobjc-framework-SceneKit==8.5
pyobjc-framework-ScreenCaptureKit==8.5
pyobjc-framework-ScreenSaver==8.5
pyobjc-framework-ScreenTime==8.5
pyobjc-framework-ScriptingBridge==8.5
pyobjc-framework-SearchKit==8.5
pyobjc-framework-Security==8.5
pyobjc-framework-SecurityFoundation==8.5
pyobjc-framework-SecurityInterface==8.5
pyobjc-framework-ServiceManagement==8.5
pyobjc-framework-ShazamKit==8.5
pyobjc-framework-Social==8.5
pyobjc-framework-SoundAnalysis==8.5
pyobjc-framework-Speech==8.5
pyobjc-framework-SpriteKit==8.5
pyobjc-framework-StoreKit==8.5
pyobjc-framework-SyncServices==8.5
pyobjc-framework-SystemConfiguration==8.5
pyobjc-framework-SystemExtensions==8.5
pyobjc-framework-UniformTypeIdentifiers==8.5
pyobjc-framework-UserNotifications==8.5
pyobjc-framework-UserNotificationsUI==8.5
pyobjc-framework-VideoSubscriberAccount==8.5
pyobjc-framework-VideoToolbox==8.5
pyobjc-framework-Virtualization==8.5
pyobjc-framework-Vision==8.5
pyobjc-framework-WebKit==8.5
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
python-lambda-local==0.1.13
pyttsx3==2.90
pyzmq==22.3.0
requests==2.27.1
s3transfer==0.6.0
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.2.0
terminado==0.15.0
tinycss2==1.1.1
tornado==6.1
traitlets==5.1.1
urllib3==1.26.9
virtualenv==20.16.5
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
